feat: implementa modulo de administrador e corrige heranca no usuario_repository

+++ src/app/repository/sqlalchemy/sqlalchemy_usuario_repository.py
@@ -13,6 +13,18 @@ class SqlAlchemyUsuarioRepository(UsuarioRepository):
         
     def find_by_email(self, email: str) -> Optional[Usuario]:
         return self.__db.query(Usuario).filter(Usuario.email == email).first()
+
+    def create(self, usuario: Usuario) -> Usuario:
+        self.__db.add(usuario)
+        self.__db.commit()
+        self.__db.refresh(usuario)
+        return usuario
+    
+    def read(self, tipo: Optional[str] = None) -> list[Usuario]:
+        query = self.__db.query(Usuario)
+        if tipo:
+            query = query.filter(Usuario.tipo == tipo)
+        return query.all()
     
     def update(self, usuario: Usuario) -> Usuario:
         usuario_atualizado = self.__db.merge(usuario)
@@ -22,5 +34,6 @@ class SqlAlchemyUsuarioRepository(UsuarioRepository):
 
     def delete(self, usuario_id: uuid.UUID) -> None: 
         usuario = self.__db.query(Usuario).filter(Usuario.id == usuario_id).first()
-        self.__db.delete(usuario)
-        self.__db.commit()
+        if usuario:
+            self.__db.delete(usuario)
+            self.__db.commit()
